Harden cURL call and prevent doubled package install
In some cases, the webserver responds with an actual HTML 404 error page instead of a raw (empty) 404 response. cURL then outputs the 404 HTML to STDOUT instead of the expected document, which may be huge or even match further processing within the calling script. It is hence good practice to use the "-f" option, to make cURL fail as well in case of a 404 HTML error page, printing a shortened error message to STDERR and leaving STDOUT empty.

apt and apt-get can be used to install package files directly, using the file path(s) as argument. The "dpkg -i" + "apt install -f" workaround to automatically install dependencies is hence not required and the irritating error output and additional processing can be prevented.

Signed-off-by: MichaIng <micha@dietpi.com>